+++ packages/mds-compliance-service/ormconfig.ts
@@ -17,4 +17,4 @@
 import { ComplianceRepository } from './repository'
 
 // Make connection options available to TypeORM CLI
-module.exports = ComplianceRepository.cli({ migrationsDir: 'repository/migrations' })
+export default ComplianceRepository.cli()