+++ packages/mds-geography-service/ormconfig.ts
@@ -17,4 +17,4 @@
 import { GeographyRepository } from './repository'
 
 // Make connection options available to TypeORM CLI
-module.exports = GeographyRepository.cli({ migrationsDir: 'repository/migrations' })
+export default GeographyRepository.cli()