+++ packages/mds-policy-service/ormconfig.ts
@@ -17,4 +17,4 @@
 import { PolicyRepository } from './repository'
 
 // Make connection options available to TypeORM CLI
-module.exports = PolicyRepository.cli({ migrationsDir: 'repository/migrations' })
+export default PolicyRepository.cli()