+++ container-images/mds-jurisdiction-service/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @container-images/mds-jurisdiction-service
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [91159748]
+- Updated dependencies [ed8ecc80]
+  - @mds-core/mds-jurisdiction-service@0.5.26
+
 ## 0.1.0
 
 ### Minor Changes