+++ packages/mds-compliance-api/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @mds-core/mds-compliance-api
 
+## 0.4.36
+
+### Patch Changes
+
+- Updated dependencies [ed8ecc80]
+  - @mds-core/mds-compliance-service@0.10.31
+  - @mds-core/mds-db@0.6.17
+  - @mds-core/mds-policy-service@0.11.28
+  - @mds-core/mds-schema-validators@0.5.10
+  - @mds-core/mds-test-data@0.4.28
+  - @mds-core/mds-api-helpers@0.2.25
+
 ## 0.4.35
 
 ### Patch Changes