+++ packages/mds-policy/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @mds-core/mds-policy
 
+## 0.7.34
+
+### Patch Changes
+
+- Updated dependencies [ed8ecc80]
+  - @mds-core/mds-db@0.6.17
+  - @mds-core/mds-policy-service@0.11.28
+  - @mds-core/mds-schema-validators@0.5.10
+  - @mds-core/mds-api-helpers@0.2.25
+
 ## 0.7.33
 
 ### Patch Changes